+++ src/lib/api/AddressApi.js
@@ -9,3 +9,13 @@ export const addressCreate = async (token, contactId, { street, city, province, 
     body: JSON.stringify({ street, city, province, country, postal_code }),
   });
 };
+
+export const addressList = async (token, contactId) => {
+  return fetch(`${import.meta.env.VITE_URL_API}/contacts/${contactId}/addresses`, {
+    method: 'GET',
+    headers: {
+      Accept: 'application/json',
+      Authorization: token,
+    },
+  });
+};